#4 refactor : 폴더 구조 변경, Optional 처리 방식 ifelse로 변경
