+++ geulkkoli/src/main/java/com/geulkkoli/domain/user/ImplUserRepository.java
@@ -1,6 +1,5 @@
-package com.geulkkoli.infrastructure;
+package com.geulkkoli.domain.user;
 
-import com.geulkkoli.domain.user.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
