--- geulkkoli/src/main/java/com/geulkkoli/infrastructure/EmptyDataException.java
@@ -1,7 +1,0 @@
-package com.geulkkoli.infrastructure;
-
-public class EmptyDataException extends RuntimeException{
-    public EmptyDataException(String message) {
-        super(message);
-    }
-}
