+++ geulkkoli/src/test/java/com/geulkkoli/domain/service/LoginTest.java
@@ -3,7 +3,7 @@ package com.geulkkoli.domain.service;
 import com.geulkkoli.domain.user.User;
 import com.geulkkoli.domain.user.service.LoginFailureException;
 import com.geulkkoli.domain.user.service.LoginService;
-import com.geulkkoli.infrastructure.UserRepository;
+import com.geulkkoli.domain.user.UserRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
