+++ geulkkoli/src/test/java/com/geulkkoli/infrastructure/PostRepositoryTest.java
@@ -1,6 +1,7 @@
 package com.geulkkoli.infrastructure;
 
 import com.geulkkoli.domain.post.Post;
+import com.geulkkoli.domain.post.PostRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -37,7 +38,7 @@ class PostRepositoryTest {
         postRepository.save(post);
 
         Post findPost = postRepository.findById(1L)
-                .orElseThrow(() -> new EmptyDataException("해당 데이터가 존제하지 않습니다."));
+                .orElse(null);
 
         assertThat(findPost).isEqualTo(post);
     }
