+++ geulkkoli/src/test/java/com/geulkkoli/infrastructure/UserRepositoryTest.java
@@ -1,6 +1,7 @@
 package com.geulkkoli.infrastructure;
 
 import com.geulkkoli.domain.user.User;
+import com.geulkkoli.domain.user.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -39,7 +40,7 @@ class UserRepositoryTest {
 
         userRepository.save(user);
         User findUser = userRepository.findById(1L)
-                .orElseThrow(() -> new EmptyDataException("해당 데이터가 존제하지 않습니다."));
+                .orElse(null);
 
         assertThat(findUser).isEqualTo(user);
     }
@@ -55,7 +56,7 @@ class UserRepositoryTest {
 
         userRepository.save(user);
         User findByLoginIdUser = userRepository.findByUserId(user.getUserId())
-                .orElseThrow(() -> new EmptyDataException("해당 데이터가 존제하지 않습니다."));
+                .orElse(null);
 
         assertThat(findByLoginIdUser).isEqualTo(user);
     }
