+++ geulkkoli/src/test/java/com/geulkkoli/respository/UserRepositoryTest.java
@@ -1,20 +1,19 @@
 package com.geulkkoli.respository;
 
 import com.geulkkoli.domain.user.User;
+import com.geulkkoli.domain.user.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
-
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 @Transactional
 class UserRepositoryTest {
     @Autowired
-    private UsersRepository usersRepository;
+    private UserRepository userRepository;
 
     @Test
     void save() {
@@ -25,9 +24,9 @@ class UserRepositoryTest {
                 .password("1234")
                 .build();
 
-        User saveUser = usersRepository.save(user);
+        User saveUser = userRepository.save(user);
 
-        assertThat(user).isEqualTo(user);
+        assertThat(saveUser).isEqualTo(user);
     }
 
     @Test
@@ -39,10 +38,11 @@ class UserRepositoryTest {
                 .password("1234")
                 .build();
 
-        usersRepository.save(user);
-        Optional<User> findUser = usersRepository.findById(1l);
+        userRepository.save(user);
+        User findUser = userRepository.findById(1L)
+                .orElse(null);
 
-        assertThat(findUser.get()).isEqualTo(user);
+        assertThat(findUser).isEqualTo(user);
 
     }
 
@@ -55,9 +55,10 @@ class UserRepositoryTest {
                 .password("1234")
                 .build();
 
-        usersRepository.save(user);
-        Optional<User> findUser = usersRepository.findByUserId("kkk");
+        userRepository.save(user);
+        User findUser = userRepository.findByUserId("kkk")
+                .orElse(null);
 
-        assertThat(findUser.get()).isEqualTo(user);
+        assertThat(findUser).isEqualTo(user);
     }
 }
